feat: Add solana-test-validator script for testing Solana network
This commit adds a new script `solana-test-validator` to the `sonic_zk_coprocessor/validator` directory. The script is used for testing the Solana network and is executed with the `cargo run` command. It accepts command line arguments and runs the Solana test validator.

Note: This commit message does not include any issue references, tags, or author names.