feat: Add sonic-printer dependency and expose remote_loader in AccountsCache
This commit adds the `sonic-printer` dependency to the `Cargo.toml` file in the `sonic_zk_coprocessor/validator` directory. It also modifies the `AccountsCache` struct in the `accounts-db/src/accounts_cache.rs` file to make the `remote_loader` field public. This change allows for using the `RemoteAccountLoader` in the `AccountsCache` struct.

Note: This commit message does not include any issue references, tags, or author names.