feat: Add Sonic BaseLayer Transaction Signature to log messages
This commit modifies the `Bank` struct in the `bank.rs` file to add the Sonic BaseLayer Transaction Signature to the log messages. It also adds a new method `post_status_to_baselayer` to post the status to the Baselayer network. This change improves the logging functionality and provides additional information about the transaction execution.

Note: This commit message does not include any issue references, tags, or author names.